cart/requirements.txt
# PostgreSQL database adapter
psycopg2-binary==2.9.9

# Additional security and utility packages
certifi==2023.11.17
urllib3==2.1.0

# Environment variable management
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0

# Type checking support
typing-extensions==4.8.0